requirements.txt
django==1.5.4
django-allauth==0.18.0
south
django-secure
django-tastypie
facebook-sdk
tweepy
